GenericQueue: extract process_batch()
